requirements.txt
Beaker==1.9.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
Chameleon==3.2
Cython==0.27.3
chardet==3.0.4
click==6.7
demjson==2.2.4
gevent==1.2.2
hupper==1.0
idna==2.6
jsonpickle==0.9.5
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mock==2.0.0
mockredispy==2.9.3
PasteDeploy==1.5.2
pbr==3.1.1
phpserialize==1.3
plaster==1.0
plaster-pastedeploy==0.4.2
Pygments==2.2.0
PyHamcrest==1.9.0
pyramid==1.9.1
pyramid-beaker==0.8
pyramid-chameleon==0.3
pyramid-debugtoolbar==4.3
pyramid-mako==1.0.2
pyramid-tm==2.2
redis==2.10.6
repoze.lru==0.7
requests==2.18.4
short-url==1.2.2
six==1.11.0
SQLAlchemy==1.1.15
transaction==2.1.2
translationstring==1.3
urllib3==1.22
venusian==1.1.0
waitress==1.1.0
WebOb==1.7.4
WebTest==2.0.29
zope.deprecation==4.3.0
zope.interface==4.4.3
zope.sqlalchemy==0.7.7
https://launchpad.net/oursql/py3k/py3k-0.9.4/+download/oursql-0.9.4.zip

